Move aoi fly to effect into Aoi layer component
to ensure it only runs when aoi changed, and not on general rerenders of the aoi control
