+++ app/scripts/components/common/map/controls/aoi/aoi-control.tsx
@@ -1,8 +1,7 @@
-import React, { useEffect, useState } from 'react';
+import React, { useState } from 'react';
 import { createPortal } from 'react-dom';
 import { Feature, Polygon } from 'geojson';
 import styled, { css } from 'styled-components';
-import bbox from '@turf/bbox';
 
 import {
   CollecticonCircleTick,
@@ -24,7 +23,6 @@ import PresetSelector from './preset-selector';
 
 import { TipToolbarIconButton } from '$components/common/tip-button';
 import { Tip } from '$components/common/tip';
-import { getZoomFromBbox } from '$components/common/map/utils';
 
 const AnalysisToolbar = styled(Toolbar)<{ visuallyDisabled: boolean }>`
   background-color: ${themeVal('color.surface')};
@@ -116,31 +114,6 @@ function AoiControl({
       }
     }
   };
-
-  useEffect(() => {
-    if (mapboxMap && aoi) {
-      /*
-      const bounds = bbox(aoi) as LngLatBoundsLike;
-      mapboxMap.fitBounds(bounds, {
-        padding: 60
-      });
-
-      Using fitBounds causes an offset in the map, so we use flyTo instead.
-      */
-
-      // Fit AOI
-      const bboxToFit = bbox(aoi);
-      const zoom = bboxToFit ? getZoomFromBbox(bboxToFit) : 14;
-      mapboxMap?.flyTo({
-        center: [
-          (bboxToFit[2] + bboxToFit[0]) / 2, // correcting the map offset by /2
-          (bboxToFit[3] + bboxToFit[1]) / 2 // correcting the map offset by /2
-        ],
-        zoom
-      });
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [aoi]); // only run on aoi change
 
   return (
     <>
